+++ app/CHANGELOG.md
@@ -3,9 +3,16 @@
 <!-- TOC -->
 
 - [Changelog](#changelog)
+- [0.2.0](#020)
 - [0.1.0](#010)
 
 <!-- TOC -->
+
+# 0.2.0
+
+- Features added:
+  - Export to TXT file the PostgreSQL users and permissions from a Cloud SQL instance
+  - Export to TXT file the PostgreSQL audit logs (INSERT, UPDATE, DELETE) from a Cloud SQL instance
 
 # 0.1.0